requirements/dev.txt
flask-restful==0.3.6
Flask-SQLAlchemy==2.3.2
pytest==3.8.1
coverage==4.5.1
python-dotenv==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.5.1
Flask-Script==0.6.6